requirements.txt
requests
pytest-asyncio

